requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
aiosqlite==0.19.0
alembic==1.12.1
redis==5.0.1
upstash-redis==0.15.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.2
pydantic==2.5.0